fix compil (patch from gentildemon / bug report from jim++, thanks)

+++ trunk/src/weapon/weapon_launcher.h
@@ -38,7 +38,7 @@ class WeaponProjectile : public PhysicalObj
     bool explode_with_collision;
     bool can_drown;
     bool camera_in_advance;
-    double begin_time;
+    uint begin_time;
 
     ExplosiveWeaponConfig& cfg;
